requirements.txt
streamlit==1.28.0
pandas==2.0.3
psycopg2-binary==2.9.7
plotly>=5.15.0
SQLAlchemy==2.0.20
pybase64==1.3.0